Move stix modeling to its own directory
