+++ yeti/core/model/stix/__init__.py
@@ -1,0 +1,2 @@
+from .sdo import StixSDO
+from .cybox import StixCYBOX
